zcbor_print.h: Move printing code to a new file zcbor_print.h
From zcbor_debug.h (deleted) and zcbor_common.h
