+++ src/zcbor_common.c
@@ -9,6 +9,7 @@
 #include <stdint.h>
 #include <string.h>
 #include "zcbor_common.h"
+#include "zcbor_print.h"
 
 _Static_assert((sizeof(size_t) == sizeof(void *)),
 	"This code needs size_t to be the same length as pointers.");
