+++ src/zcbor_encode.c
@@ -10,6 +10,7 @@
 #include <string.h>
 #include "zcbor_encode.h"
 #include "zcbor_common.h"
+#include "zcbor_print.h"
 
 _Static_assert((sizeof(size_t) == sizeof(void *)),
 	"This code needs size_t to be the same length as pointers.");
