+++ tests/decode/test3_simple/src/main.c
@@ -7,7 +7,6 @@
 #include <zephyr/ztest.h>
 #include "pet_decode.h"
 #include "serial/serial_recovery_decode.h"
-#include "zcbor_debug.h" // Enables use of print functions when debugging tests.
 
 uint8_t serial_rec_input1[] = {
 	/* "data" */
