+++ tests/decode/test5_corner_cases/src/main.c
@@ -8,7 +8,6 @@
 #include <math.h>
 #include <corner_cases.h>
 #include <zcbor_decode.h>
-#include <zcbor_debug.h> // Enables use of print functions when debugging tests.
 
 #define CONCAT_BYTE(a,b) a ## b
