+++ tests/encode/test3_corner_cases/src/main.c
@@ -6,7 +6,6 @@
 
 #include <zephyr/ztest.h>
 #include "corner_cases_encode.h"
-#include "zcbor_debug.h" // Enables use of print functions when debugging tests.
 
 #define CONCAT_BYTE(a,b) a ## b
