+++ tests/unit/test1_unit_tests/src/main.c
@@ -7,7 +7,7 @@
 #include <zephyr/ztest.h>
 #include "zcbor_decode.h"
 #include "zcbor_encode.h"
-#include "zcbor_debug.h"
+#include "zcbor_print.h"
 
 ZTEST(zcbor_unit_tests, test_int64)
 {
